+++ CHANGELOG.md
@@ -6,11 +6,13 @@
 
 ### Added
 - Experimental support for USB connectivity on SPIKE Prime ([pybricks-micropython#208]).
+- Initial support for `pybricks.iodevices.UARTDevice` ([support#220]). 
 
 ### Changed
 - Extensive overhaul of UART and port drivers on all hubs. This affects all
   official LEGO sensors on all hubs.
 
+[support#220]: https://github.com/pybricks/support/issues/220
 [pybricks-micropython#208]: https://github.com/pybricks/pybricks-micropython/pull/208
 
 ## [3.6.1] - 2025-03-11